+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on and uses the types of changes according to [Keep a Change
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- SqlServerDsc
+  - Added Test-SqlDscIsRole to be used like Test-SqlDscIsLogin but tests for a server role as principal
 
 ### Removed